backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
sqlalchemy==1.4.17
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
